pyproject.toml
[tool.poetry]
name = "slack-get-history"
version = "0.1.0"
description = ""
authors = ["shin5ok <shin5ok@55mp.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
langchain = "^0.0.271"
google-cloud-aiplatform = "^1.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
